refactor: modified imports from create-activity
