+++ src/participating-unit/participating-unit.module.ts
@@ -1,0 +1,4 @@
+import { Module } from '@nestjs/common';
+
+@Module({})
+export class ParticipatingUnitModule {}
